Move writing of a comments (VML) part to a dedicated writer. Better encapsulation, more modular.
